feature: Add language switch button
